requirements.txt
# requirements.txt
numpy>=1.20.0
scikit-learn>=1.0.0
plotly>=5.0.0
pillow>=8.0.0
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.900
